[infra-proxy-service] Update go-chef module with chef forked repo      - https://github.com/chef/go-chef      - minor update regarding server delete the message      - constructing the base URL based on fqdn|ipAddress value
