+++ .bldr.toml
@@ -1996,9 +1996,9 @@ paths = [
   "lib/uuid4/*",
   "lib/version/*",
   "vendor/github.com/apache/thrift/*",
+  "vendor/github.com/chef/go-chef/*",
   "vendor/github.com/codahale/hdrhistogram/*",
   "vendor/github.com/fsnotify/fsnotify/*",
-  "vendor/github.com/go-chef/chef/*",
   "vendor/github.com/gofrs/uuid/*",
   "vendor/github.com/golang-migrate/migrate/*",
   "vendor/github.com/golang/protobuf/*",
