+++ components/infra-proxy-service/server/v1/proxy.go
@@ -1,0 +1,98 @@
+package v1
+
+import (
+	"context"
+	"net/url"
+
+	uuid "github.com/chef/automate/lib/uuid4"
+	chef "github.com/chef/go-chef"
+	"github.com/pkg/errors"
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
+
+	"github.com/chef/automate/components/infra-proxy-service/service"
+)
+
+// ChefConfig is a V1 infra-proxy server
+type ChefConfig struct {
+	Name    string
+	Key     string
+	SkipSSL bool
+	BaseURL string
+}
+
+// NewChefClient is a V1 infra-proxy server
+func NewChefClient(config *ChefConfig) (*chef.Client, error) {
+
+	// build a client
+	client, err := chef.NewClient(&chef.Config{
+		Name:    config.Name,
+		Key:     config.Key,
+		SkipSSL: config.SkipSSL,
+		BaseURL: config.BaseURL,
+	})
+
+	if err != nil {
+		return nil, errors.Wrap(err, err.Error())
+	}
+
+	return client, nil
+}
+
+func (s *Server) createClient(ctx context.Context, orgID string) (*chef.Client, error) {
+	ctx, cancel := context.WithCancel(ctx)
+	defer cancel()
+
+	UUID, err := uuid.FromString(orgID)
+	if err != nil {
+		return nil, status.Errorf(codes.InvalidArgument, "invalid org id: %s", err.Error())
+	}
+
+	// TODO: Call GetOrgByName in order to avoid separate query to fetch infra server detail
+	// Prefer LEFT OUTER join to in order to append server detail in response
+	org, err := s.service.Storage.GetOrg(ctx, UUID)
+	if err != nil {
+		return nil, service.ParseStorageError(err, orgID, "org")
+	}
+
+	ServerID, err := uuid.FromString(org.ServerId)
+	if err != nil {
+		return nil, status.Errorf(codes.InvalidArgument, "invalid server: %s", err.Error())
+	}
+
+	server, err := s.service.Storage.GetServer(ctx, ServerID)
+	if err != nil {
+		return nil, service.ParseStorageError(err, ServerID, "org")
+	}
+
+	baseURL, err := targetURL(server.Fqdn, server.IpAddress, org.Name)
+	if err != nil {
+		return nil, status.Errorf(codes.InvalidArgument, "invalid server url: %s", baseURL)
+	}
+
+	client, err := NewChefClient(&ChefConfig{
+		Name:    org.AdminUser,
+		Key:     org.AdminKey,
+		SkipSSL: true,
+		BaseURL: baseURL,
+	})
+
+	return client, err
+}
+
+// targetURL is constructing the base URL based on fqdn|ipAddress value
+func targetURL(fqdn string, IPAddress string, orgName string) (string, error) {
+	path := fqdn
+	if path == "" {
+		path = IPAddress
+	}
+	path = path + "/organizations/" + orgName + "/"
+
+	baseURL, err := url.Parse(path)
+	if err != nil {
+		return "", errors.Wrap(err, err.Error())
+	}
+	baseURL.Scheme = "https"
+
+	return baseURL.String(), nil
+}
